feat(harness): lift WebhookConfig into culture/clients/shared/webhook_types
The dataclass was byte-identical across all four backends. Lifting it into
the shared tier kills 4 copies and prepares webhook.py to import from
shared next. Each backend's config.py keeps a one-line re-export so existing
`from culture.clients.<backend>.config import WebhookConfig` callers
keep working unchanged.

Refs #357.
